Previous change broke other PWM devices.
The changes were only tested with a _servo_ device, which only reacts to a single pulse to activate, versus something like an LED which requires the _constant_ inputs to keep the value.

On a hunch, replaced the lock-based sleep with the newer busy-wait and the previously observed jitter on the servo is almost entirely eliminated. Also tested with an LED, which also now does not exhibit a previously observed flickering (I thought I had lousy LED's :shrug:).

Bumped the log message to WARN so it's a bit more obvious that what is being used may not be what the user wanted. :smiley: