Stream modifyes the underlying map while traversing.
Although it's not inherently obvious (even though the comment indicated it), removing elements from the underlying collection while in a stream is technically Not A Good Idea(tm). The change here collects the elements in a new list and **then** invokes the `close` method, which _should_ invoke the `closed` method on this class.